+++ htable.h
@@ -19,14 +19,15 @@
 
 typedef struct hashtable T_HashTable;
 typedef void *T_HashTableEl;
-typedef unsigned int (*T_HashFunc)(void *);
+typedef size_t (*T_HashFunc)(void *);
 typedef int (*T_ComparFunc)(void *, void *);
 
 
-int make_ht(T_HashFunc f1, T_HashFunc f2, T_ComparFunc c, int size, T_HashTable **H);
-int hash_add(T_HashTable *H, T_HashTableEl *E, int *hint);
-int hash_remove(T_HashTable *H, T_HashTableEl *E, int hint);
+int make_ht(T_HashFunc f1, T_HashFunc f2, T_ComparFunc c, size_t size,
+	    T_HashTable **H);
+int hash_add(T_HashTable *H, T_HashTableEl *E, size_t *hint);
+int hash_remove(T_HashTable *H, T_HashTableEl *E, size_t hint);
 int hash_lookup(T_HashTable *H, T_HashTableEl *E, T_HashTableEl **E2,
-		int *hint);
+		size_t *hint);
 int free_ht(T_HashTable *H, T_HashFunc entry_free);
